requirements.txt
# Install PyTorch with CUDA 12.6 support (RTX 4070 / any Nvidia GPU)
# If you don't have a GPU, replace cu126 with cpu:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cu126
torchvision==0.26.0+cu126

# Web framework
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.24
starlette==1.0.0

# Database
sqlalchemy==2.0.49
aiosqlite==0.22.1
greenlet==3.3.2

# ML / data
numpy==2.4.3
pillow==12.1.1
scikit-learn==1.8.0
matplotlib==3.10.8
pandas==3.0.2
tqdm==4.67.3

# Pydantic / validation
pydantic==2.12.5
annotated-types==0.7.0
